pyproject.toml
[project]
name = "core-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.9.5",
    "asyncmy==0.2.8",
    "fastapi==0.100.1",
    "pandas>=2.0.3",
    "python-dotenv==1.0.0",
    "sqlalchemy==2.0.19",
    "starlette-prometheus==0.9.0",
    "uvicorn==0.23.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.3",
    "ruff>=0.8.1",
]
